# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Navigation Api",
        "description": "Navigation List Api",
        "version": "1.0"
    },
    "paths": {
        "/api/add-location": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Add a new location",
                "description": "Add a new location\n\nThis endpoint allows you to add a new location by providing latitude, longitude, name, and marker color.",
                "operationId": "4b4f7a6e87e713637d56b60a78ac452d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "example": 40.748817
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": -73.985428
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Sample Location"
                                    },
                                    "marker_color": {
                                        "type": "string",
                                        "example": "#111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Location added successfully"
                                        },
                                        "location": {
                                            "$ref": "#/components/schemas/Location"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "442": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "latitude": [
                                                    "The latitude field is required."
                                                ],
                                                "longitude": [
                                                    "The longitude field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/location/{id}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get Location Detail",
                "operationId": "5aa0863a0a475f7060132f62f8598ac4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/locations": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get all locations with pagination.",
                "description": "Get a paginated list of all locations.",
                "operationId": "7bff142cb4941544ae2add489283a2b1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page.",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "limit": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sorted-locations": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Get all locations sorted by distance.",
                "description": "Get a list of all locations sorted by distance from a specified point.",
                "operationId": "184c7131d05133f51197e9a6853fe814",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40.748817
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -73.985428
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "442": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "latitude": [
                                                    "The latitude field is required."
                                                ],
                                                "longitude": [
                                                    "The longitude field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Location": {
                "title": "Location",
                "description": "Location model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}